♻️ refactor: 세로형 카드로 분리
